Breakout SwitchDrawable
